Remove PageCreator from MwIntegrationTestCase
